fix(mcp): remove unnecessary i64 cast in cleanup_sync_data handler
Clippy 1.93.0 (CI) now flags identity casts (i64 -> i64).
The `deleted` variable is already i64 from cleanup_sync_data().

+++ src/mcp/handlers.rs
@@ -2144,7 +2144,7 @@ impl ToolHandler {
             tracing::warn!("Failed to clean Meilisearch code index: {}", e);
         }
         Ok(
-            json!({"deleted_count": deleted as i64, "message": "Sync data cleaned. Run sync_project to rebuild."}),
+            json!({"deleted_count": deleted, "message": "Sync data cleaned. Run sync_project to rebuild."}),
         )
     }
